+++ changelog.md
@@ -1,3 +1,7 @@
+Camera Plugin 0.2.1 (04.11.2013)
+---------------------------
+Experimental Release for CryEngine 3.5.4 (32/64 bit)
+
 Camera Plugin 0.2 (27.04.2013)
 ---------------------------
 Experimental Release for CryEngine 3.4.5 (64 bit)